Add spread comments, round-trip tests, and fix stale doc
- Clarify DFA/NFA spread limitation in dfaMatcher.ts and environment.ts

- Add round-trip tests for object spread in grammarRuleWriter.spec.ts

- Update stale doc comment in environment.ts for CompiledObjectElement[]
